requirements.txt
aws-cdk-lib==2.186.0
constructs>=10.0.0,<11.0.0
pytest==6.2.5
boto3>=1.26.0
requests>=2.31.0
